Duo

Documentation for Duo.

Duo.Cell2D

Duo.Point2D

Base.contains

Duo.children

Duo.divide!

Duo.extents

Duo.has_children

Duo.is_root

Duo.parent

Duo.Cell2D — Type

Cell2D

A square shape domain centered at Point2D with side length of size, equal to the square's width and height.

The level is the number of cell divisions that have occurred to create the subject cell; root cells by definition have level = 0.

The parent is the cell that sits above the current subject cell in the hierarchy. A root cell has no parent by definition. A non-root (child) cell must have a non-nothing parent.

The children are either nothing or the four children cells sitting below the current subject cell in the hierarchy.

Duo.Point2D — Type

Point2D

A coordinate with x and y values measured from an origin.

Base.contains — Method

contains(cell, point)

Determines if the coordinates of a point lie on or within the boundary of a cell`.

Returns true if the point is on the boundary of the cell or in the interior of the cell; false otherwise.

Duo.children — Method

children(cell)

Get the children (cells) of the current subject cell cell.

Prior to cell division, a cell has no children. After cell division in 2D, the cell has exactly four children. After cell division in 3D, the cell has exactly eight children.

Duo.divide! — Method

divide!(cell)

Causes the cell to undergo cell division. In 2D, causes the current subject cell to have four children.

Duo.extents — Method

extents(cell)

In 2D, returns the left/lower corner (xmin, ymin) and the right/upper corner (xmax, ymax) as a tuple ((xmin, ymin), (xmax, ymax)).

Duo.has_children — Method

has_children(cell)

Whether the subject cell cell has children or not. Returns true if the cell has children; false if the cell has no children.

Duo.is_root — Method

is_root(cell)

Whether or not the subject cell cell is the root of the cell hierarchy. Returns true if the cell is the root; false otherwise.

Duo.parent — Method

parent(cell)

Get the parent cell of the current subject cell cell.